# packages/contracts/openapi/live-match-impact.openapi.yaml
openapi: 3.1.0
info:
  title: Live Match Impact API
  version: 1.1.0
  description: Public client-facing API for football intelligence consumers such as the Chrome extension and OBS overlays.
servers:
  - url: https://example.com
paths:
...
  /billing/status:
    get:
      operationId: getBillingStatus
      parameters:
        - in: query
          name: user_id
          schema:
            type: string
      responses:
        "200":
          description: Current billing status
          content:
            application/json:
              schema:
                $ref: ../schemas/billing-status.json
...
  /matches/upcoming:
    get:
      operationId: getUpcomingMatches
      parameters:
        - in: query
          name: date
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: number
      responses:
        "200":
          description: Upcoming matches
          content:
            application/json:
              schema:
                $ref: ../schemas/upcoming-matches.json
  /match-impact:
    get:
      operationId: getMatchImpact
      parameters:
        - in: query
          name: fixture_id
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Match impact payload
          content:
            application/json:
              schema:
                $ref: ../schemas/match-impact.json
...